+++ Compra.cs
@@ -11,7 +11,7 @@ namespace Slc_Mercado
         public int id { get; set; }
         public Usuario comprador { get; set; }
         public Dictionary <Producto,int> productos;
-        public double total { get; set; }
+        public decimal total { get; set; }
 
         public Compra() { }
         public Compra(int id,Usuario comprador, Dictionary<Producto,int> productosCarrito) {
@@ -31,9 +31,9 @@ namespace Slc_Mercado
             return "ID: "+ this.id + "- Usuario: " + this.comprador.ToString() + "- Productos: " + this.productos.ToString() + "- Total: " + this.total;
         }
 
-        internal double calcularCompra()
+        internal decimal calcularCompra()
         {
-            double suma = 0;
+            decimal suma = 0;
             foreach (KeyValuePair<Producto, int> kvp in productos)
             {
                 suma += kvp.Key.precio * kvp.Value; //revisar getter
